Update gameboard with current players token
This commit reflects a large change to the codebase. I initially wrote
the board update function to check if the move passed in to it was a falsy
value (null) and if so, dynamically access the board array based on the
array index number passed in as an argument and assign the current player's
token. I was unable to update the local board array data model variable
inside of the method found within the function.

To accomplish this functionality I refactored the initializeGameboard
function to return an object with a local data model variable, with
methods to update that variable using the 'this' keyword.

The gameboard is updated using the current player's token but I noticed
that when the switchCurrentPlayer function was invoked, the getCurrentPlayer
method called in the updateBoard method was no longer being successfully
invoked. I was previously tracking the current player simply by the current
player's number attributed to it. I created an array that contained both
player objects then refactored the switchCurrentPlayer functionality to
switch back and forth between those two objects.

Lastly I refactored the trackCurrentPlayer function to return an object
with methods as I did for the initializeGameboard function.