Account for more than 3 consecutive winning moves
Initially, the checkForWin function was only accounting for if the user
made three consecutive winning moves. This is not going to be the case
every time the function is called.

I refactored the function to loop through each of the win conditions
and pass each winCondition into a separate function that loops through
the current indices of the gameboard that the current player's tokens
occupy and checks if the winCondition contains that index number. If
it does, that index number is assigned to an array. If, after the for
loop finishes, the array is holding 3 numbers, a winning combination
has been found.

That winning condition is then assigned to a new global variable that
will be used to highlight the cells in the tic tac toe grid that pertain
to the winning combo.