Alternate starting player no matter the winner
In this commit, I started by storing the global variables in a single
object called 'store'. I declared a new global variable called starting
player which is initialized to player one on window load. When the game
is reset, the current player is reassigned to the opposite of what the
starting player is assigned to, then the starting player is reassigned
to reflect the current player as the new game begins.